Added command cb. Working, but not done.

+++ enioka_scan/src/main/java/com/enioka/scanner/sdk/zebraoss/SsiMessage.java
@@ -78,7 +78,7 @@ public enum SsiMessage {
         this.parser = parser;
     }
 
-    public int getValue() {
+    public byte getValue() {
         return opCode;
     }
